requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==3.0.10
requests==2.26.0
openai>=1.0.0
python-dotenv==0.19.0
newspaper3k==0.2.8
elevenlabs==0.2.12
boto3==1.26.0
httpx>=0.24.1